refactor.

+++ Haskell/quick-sort/src/Lib.hs
@@ -12,7 +12,7 @@ quickSort (x:xs) = quickSort left ++ [x] ++ quickSort right
           right = filter (>= x) xs
 
 generateRandomList :: Int -> IO [Int]
-generateRandomList n = sequence $ replicate n $ randomRIO (1,100::Int)
+generateRandomList n = sequence $ replicate n $ randomRIO (1, 100)
 
 someFunc :: IO ()
 someFunc = (quickSort <$> generateRandomList 100) >>= print
